requirements.txt
backoff==2.2.1
numpy==2.2.5
pandas==2.2.3
pytest==8.3.3
python-dotenv==1.1.0
Requests==2.32.3
torch==2.6.0
psycopg2-binary
torch
scikit-learn